Refactor: Consolidate project structure and build system
This commit significantly refactors the Alte text editor project to improve organization, clarity, and maintainability.

Key changes include:

1.  **Unified Project Structure:**
    *   Removed the redundant `AlteEditor/` subdirectory.
    *   Consolidated all source code into the top-level `src/` (implementation files) and `include/` (header files) directories.
    *   Prioritized and integrated more recent code components (like `AlteRope`, `AlteThemeManager`, `splashscreen`) from the former `AlteEditor/` directory.
    *   Standardized file naming (e.g., `AlteSyntaxHighlighter.cpp`).

2.  **Consolidated CMake Build System:**
    *   The top-level `CMakeLists.txt` is now the single authoritative build configuration.
    *   Upgraded project to C++20.
    *   Implemented Qt6 support with a fallback to Qt5.
    *   Ensured all source files and necessary headers (for MOC) are included in the build.
    *   The output executable is named `Alte`.

3.  **Resource Management:**
    *   All resources (icons, themes, syntax definitions) are managed under the top-level `resources/` directory.
    *   Created `resources/Alte.qrc` to embed the application icon.
    *   CPack configuration in `CMakeLists.txt` handles installation of resources.

4.  **Documentation Refactoring:**
    *   The main `README.md` has been rewritten to be a concise project overview, including build and installation instructions.
    *   The detailed project vision, manifesto, and original brainstorming notes have been moved to `VISION.md`.

5.  **Build/Environment Cleanup:**
    *   Updated `.gitignore` to exclude build output directories (e.g., `build_temp/`, `build_*/`) and common IDE user files (`*.user`).

These changes establish a cleaner foundation for future development of the Alte text editor, making it easier to navigate, build, and contribute to the project.
