--- AlteEditor/Hadaf.txt
@@ -1,1 +1,0 @@
-هدف ما چیست : بهترین سبک ترین و جالب ترین و زیبا ترین اپلیکیشن متن ادیتور ساده برای انواع فرمت بسیار هوشمند و وو و و و
