+++ include/AlteThemeManager.h
