+++ resources/Alte.qrc
@@ -1,0 +1,5 @@
+<RCC>
+    <qresource prefix="/">
+        <file alias="icons/alte_icon.png">icons/alte_icon.png</file>
+    </qresource>
+</RCC>
