+++ src/AlteRope.cpp
