+++ src/AlteSyntaxHighlighter.cpp
@@ -1,242 +1,220 @@
-#include "AlteSyntaxHighlighter.h"
-#include "AlteTheme.h" // Required for AlteTheme
-#include <QFile>
-#include <QJsonDocument>
-#include <QJsonObject>
+#include "syntaxhighlighter.h"
+#include "AlteThemeManager.h"
+#include <QTextDocument>
 #include <QJsonArray>
 #include <QDebug>
 
-AlteSyntaxHighlighter::AlteSyntaxHighlighter(QTextDocument *parent, AlteTheme *theme)
-    : QSyntaxHighlighter(parent), m_theme(theme) {
-    if (!m_theme) {
-        qWarning() << "AlteSyntaxHighlighter initialized with a null theme!";
+SyntaxHighlighter::SyntaxHighlighter(QTextDocument *parent, AlteThemeManager *themeManager, const QString& languageName)
+    : QSyntaxHighlighter(parent) {
+    if (themeManager && !languageName.isEmpty()) {
+        setCurrentLanguage(languageName, themeManager);
+    } else {
+        qWarning() << "SyntaxHighlighter: ThemeManager or languageName not provided. No rules loaded.";
     }
 }
 
-void AlteSyntaxHighlighter::loadRulesFromLanguageFile(const QString &filePath) {
-    m_rules.clear(); // Use m_rules
-
-    if (filePath.isEmpty() || !m_theme) {
-        if (!m_theme) qWarning() << "No theme set in highlighter, cannot apply formats.";
-        if (filePath.isEmpty()) qInfo() << "No language file provided, clearing rules.";
+void SyntaxHighlighter::setCurrentLanguage(const QString& languageName, AlteThemeManager *themeManager) {
+    m_highlightingRules.clear();
+    if (!themeManager) {
+        qWarning() << "SyntaxHighlighter::setCurrentLanguage: ThemeManager is null.";
         rehighlight();
         return;
     }
-
-    QFile langFile(filePath);
-    if (!langFile.exists()) {
-        qWarning() << "Language file does not exist:" << filePath;
+    if (languageName.isEmpty()){
+        qWarning() << "SyntaxHighlighter::setCurrentLanguage: languageName is empty.";
         rehighlight();
         return;
     }
-    if (!langFile.open(QIODevice::ReadOnly)) {
-        qWarning() << "Could not open language file:" << filePath;
-        rehighlight();
-        return;
+    loadRulesForLanguage(languageName, themeManager);
+    rehighlight(); // Trigger re-highlighting of the entire document
+}
+
+QTextCharFormat SyntaxHighlighter::createFormatFromRule(const QJsonObject& ruleDetails,
+                                                      const QJsonObject& themeColors, // Not used directly, themeManager is used
+                                                      const QFont& defaultFont,
+                                                      AlteThemeManager* themeManager) {
+    QTextCharFormat format;
+    QString colorName = ruleDetails.value("color").toString();
+    if (!colorName.isEmpty()) {
+        QColor color = themeManager->getColor(colorName); // getColor can lookup from "colors" or parse hex
+        if (!color.isValid()) { // Fallback if color name is not in theme's "colors" and not a valid hex
+             qWarning() << "SyntaxHighlighter: Color name/hex '" << colorName << "' is invalid or not found in theme colors. Using default text color.";
+             color = themeManager->getColor("text", Qt::black);
+        }
+        format.setForeground(color);
     }
 
-    QByteArray jsonData = langFile.readAll();
-    langFile.close();
-    QJsonParseError error;
-    QJsonDocument doc = QJsonDocument::fromJson(jsonData, &error);
-
-    if (doc.isNull()) {
-        qWarning() << "Failed to parse language JSON from" << filePath << ":" << error.errorString();
-        rehighlight();
-        return;
+    if (ruleDetails.value("bold").toBool(false)) {
+        format.setFontWeight(QFont::Bold);
+    }
+    if (ruleDetails.value("italic").toBool(false)) {
+        format.setFontItalic(true);
     }
 
-    QJsonObject rootObj = doc.object();
-    QJsonArray rulesArray = rootObj["highlighting_rules"].toArray();
-
-    for (const QJsonValue &val : rulesArray) {
-        QJsonObject ruleObj = val.toObject();
-        QString syntaxType = ruleObj["type"].toString();
-        QString styleKey = ruleObj["style_key"].toString(syntaxType); // Fallback to syntaxType if style_key missing
-
-        // Map syntax type from language file to theme format key if necessary (as before)
-        // This can be made more generic or data-driven.
-        if (syntaxType == "keywords") styleKey = "keyword";
-        else if (syntaxType == "line_comment") styleKey = "comment";
-        else if (syntaxType == "strings_double_quotes" || syntaxType == "strings_single_quotes") styleKey = "string";
-        else if (syntaxType == "numbers") styleKey = "number";
-        // For "multi_line_comment", styleKey would ideally be "comment" or "block_comment"
-        // The JSON should provide this directly via "style_key". If it's "multi_line_comment",
-        // and style_key is "comment", it will use the comment theme.
-
-        QTextCharFormat format = m_theme->getFormat(styleKey);
-
-        if (syntaxType == "multi_line_comment" || syntaxType == "multi_line_string") {
-            QString startPatternStr = ruleObj["start_pattern"].toString();
-            QString endPatternStr = ruleObj["end_pattern"].toString();
-            int stateId = ruleObj["state_id"].toInt(0);
-
-            if (startPatternStr.isEmpty() || endPatternStr.isEmpty() || stateId == 0) {
-                qWarning() << syntaxType << "rule is missing start_pattern, end_pattern, or has invalid state_id (0). Name:" << ruleObj["name"].toString();
-                continue;
+    if (ruleDetails.contains("fontPointSizeOffset")) {
+        int offset = ruleDetails.value("fontPointSizeOffset").toInt(0);
+        if (offset != 0) {
+            QFont currentFont = format.font(); // Get font if already modified by bold/italic
+            if (currentFont.pointSize() <= 0) { // If not set, use default
+                 currentFont = defaultFont;
             }
-            // For multi-line strings, QRegularExpression::DotMatchesEverythingOption might be important
-            // if the string content itself can span newlines and the regex needs to match across them.
-            // However, QSyntaxHighlighter processes block by block (line by line).
-            // The state mechanism handles the "spanning newlines" part.
-            // The patterns themselves usually match within a single line.
-            // For example, """ might be a start/end pattern.
-            QRegularExpression::PatternOptions options = QRegularExpression::NoPatternOption;
-            if (ruleObj.contains("options") && ruleObj["options"].toString() == "DotMatchesEverythingOption") {
-                 options = QRegularExpression::DotMatchesEverythingOption;
-            }
-
-            m_rules.append(HighlightingRule(QRegularExpression(startPatternStr, options),
-                                            QRegularExpression(endPatternStr, options),
-                                            format, stateId, styleKey));
-        } else if (ruleObj.contains("list")) { // Keywords
-            QStringList keywords = ruleObj["list"].toVariant().toStringList();
-            for (const QString &keyword : keywords) {
-                m_rules.append(HighlightingRule(QRegularExpression(QString("\\b%1\\b").arg(keyword)), format, styleKey));
-            }
-        } else if (ruleObj.contains("start_delimiter") && ruleObj.contains("end_delimiter")) { // Strings
-            QString start = QRegularExpression::escape(ruleObj["start_delimiter"].toString());
-            QString end = QRegularExpression::escape(ruleObj["end_delimiter"].toString());
-            m_rules.append(HighlightingRule(QRegularExpression(QString("%1[^%2\\n]*%2").arg(start).arg(end)), format, styleKey));
-        } else if (ruleObj.contains("start_delimiter")) { // Line comments
-            QString start = QRegularExpression::escape(ruleObj["start_delimiter"].toString());
-            m_rules.append(HighlightingRule(QRegularExpression(QString("%1[^\n]*").arg(start)), format, styleKey));
-        } else if (syntaxType == "number" && ruleObj.contains("pattern")) { // Number rule
-            QString patternStr = ruleObj["pattern"].toString();
-            if (!patternStr.isEmpty()) {
-                 QRegularExpression::PatternOptions options = QRegularExpression::NoPatternOption;
-                 if (ruleObj.contains("options") && ruleObj["options"].toString() == "CaseInsensitiveOption") {
-                    // Example if numbers could have case-specific parts like '0xFF' vs '0XFF'
-                    options = QRegularExpression::CaseInsensitiveOption;
-                 }
-                m_rules.append(HighlightingRule(QRegularExpression(patternStr, options), format, styleKey));
+            int newSize = currentFont.pointSize() + offset;
+            if (newSize > 0) {
+                currentFont.setPointSize(newSize);
+                format.setFont(currentFont); // Apply the font with new size
             } else {
-                qWarning() << "Number rule is missing pattern. Name:" << ruleObj["name"].toString();
+                qWarning() << "SyntaxHighlighter: Calculated font point size is not positive (" << newSize << "). Ignoring offset.";
             }
         }
-         // Add more rule parsing logic here as needed
     }
-
-    qDebug() << "Loaded" << m_rules.count() << "rules from" << filePath;
-    rehighlight();
+    return format;
 }
 
-void AlteSyntaxHighlighter::highlightBlock(const QString &text) {
-    if (!m_theme) return;
-
-    // --- Handle multi-line state continuation ---
-    int previousStateId = previousBlockState();
-    const HighlightingRule* activeMultiLineRule = nullptr;
-
-    if (previousStateId > 0) { // Previous block was part of a multi-line construct
-        activeMultiLineRule = findMultiLineRuleByStateId(previousStateId);
-        if (activeMultiLineRule) {
-            QRegularExpressionMatch endMatch = activeMultiLineRule->endPattern.match(text);
-            int endIndex = endMatch.capturedStart();
-            int endLength = endMatch.capturedLength();
-
-            if (endIndex != -1) { // End pattern found in current block
-                setFormat(0, endIndex + endLength, activeMultiLineRule->format);
-                setCurrentBlockState(0); // Return to default state
-            } else { // End pattern not found, whole block is part of the multi-line comment
-                setFormat(0, text.length(), activeMultiLineRule->format);
-                setCurrentBlockState(activeMultiLineRule->stateId);
-                return; // Block fully processed as multi-line
-            }
-        } else {
-             // Invalid previous state, reset
-            setCurrentBlockState(0);
-        }
+void SyntaxHighlighter::loadRulesForLanguage(const QString& languageName, AlteThemeManager *themeManager) {
+    m_highlightingRules.clear();
+    QJsonObject langRules = themeManager->getSyntaxRulesForLanguage(languageName);
+    if (langRules.isEmpty()) {
+        qWarning() << "SyntaxHighlighter: No syntax rules found for language" << languageName;
+        return;
     }
 
-    // --- Process single-line rules and new multi-line starts ---
-    // This part needs to be careful about rule order and overlaps.
-    // A common strategy: apply multi-line first, then single-line for remaining parts.
-    // Or, check single-line, but if a multi-line starts, it takes precedence for the rest of the block.
+    QFont documentFont;
+    if (document()) {
+        documentFont = document()->defaultFont();
+    } else {
+        documentFont = QApplication::font(); // Fallback, should ideally not happen
+        qWarning() << "SyntaxHighlighter: QTextDocument is null during rule loading. Using application default font.";
+    }
 
-    int searchStartIndex = 0; // Current position in the text to start searching for rules
+    // The themeColors parameter in createFormatFromRule is currently unused,
+    // as themeManager->getColor() is preferred for direct lookup or hex parsing.
+    // Passing an empty QJsonObject for now.
+    QJsonObject dummyColors;
 
-    while(searchStartIndex < text.length()) {
-        int blockStateToSet = 0; // Default state for end of block
-        int earliestMatchStart = text.length();
-        const HighlightingRule* matchedRuleThisPass = nullptr;
-        int matchedLength = 0;
+    for (const QString& ruleKey : langRules.keys()) {
+        QJsonObject ruleDef = langRules.value(ruleKey).toObject();
+        HighlightingRule currentRuleSetup; // For properties common to a set of patterns
+        currentRuleSetup.format = createFormatFromRule(ruleDef, dummyColors, documentFont, themeManager);
+        currentRuleSetup.isBlockRule = ruleDef.value("block").toBool(false);
 
-        // Check for the start of any multi-line rule FIRST
-        for (const HighlightingRule &rule : m_rules) {
-            if (rule.type != RuleType::MultiLine) continue;
-
-            QRegularExpressionMatch match = rule.pattern.match(text, searchStartIndex);
-            if (match.hasMatch() && match.capturedStart() < earliestMatchStart) {
-                earliestMatchStart = match.capturedStart();
-                matchedRuleThisPass = &rule; // Potential multi-line start
-                // Matched length for a multi-line start is just the start pattern itself initially
-                // The actual highlighting will depend on finding the end or EOB
+        if (currentRuleSetup.isBlockRule) {
+            currentRuleSetup.pattern = QRegularExpression(ruleDef.value("start_pattern").toString());
+            currentRuleSetup.endPattern = QRegularExpression(ruleDef.value("end_pattern").toString());
+            if (currentRuleSetup.pattern.isValid() && currentRuleSetup.endPattern.isValid()) {
+                m_highlightingRules.append(currentRuleSetup);
+            } else {
+                qWarning() << "SyntaxHighlighter: Invalid regex for block rule" << ruleKey
+                           << ": Start:" << ruleDef.value("start_pattern").toString()
+                           << "End:" << ruleDef.value("end_pattern").toString();
             }
-        }
-
-        // If a multi-line rule starts before any single-line rule found so far (or if no single yet)
-        if (matchedRuleThisPass) { // A multi-line rule start was found
-            int commentStartIndex = earliestMatchStart;
-            QRegularExpressionMatch endMatch = matchedRuleThisPass->endPattern.match(text, commentStartIndex + matchedRuleThisPass->pattern.match(text, commentStartIndex).capturedLength());
-
-            if (endMatch.hasMatch()) { // End is in the same block
-                int commentEndIndex = endMatch.capturedStart() + endMatch.capturedLength();
-                setFormat(commentStartIndex, commentEndIndex - commentStartIndex, matchedRuleThisPass->format);
-                searchStartIndex = commentEndIndex; // Continue after this multi-line comment
-                blockStateToSet = 0; // Back to default state
-            } else { // End is not in this block
-                setFormat(commentStartIndex, text.length() - commentStartIndex, matchedRuleThisPass->format);
-                blockStateToSet = matchedRuleThisPass->stateId;
-                searchStartIndex = text.length(); // Whole block consumed
+        } else if (ruleDef.contains("pattern")) { // Single regex pattern
+            currentRuleSetup.pattern = QRegularExpression(ruleDef.value("pattern").toString());
+             if (currentRuleSetup.pattern.isValid()) {
+                m_highlightingRules.append(currentRuleSetup);
+            } else {
+                qWarning() << "SyntaxHighlighter: Invalid regex pattern for rule" << ruleKey << ":" << ruleDef.value("pattern").toString();
             }
-            setCurrentBlockState(blockStateToSet);
-            if(searchStartIndex >= text.length()) continue; // Move to next block if fully consumed
-        }
-
-        // If no multi-line rule started, or we are past a completed multi-line segment, process single-line rules
-        // Reset earliestMatch for single-line rules for the current searchStartIndex
-        earliestMatchStart = text.length();
-        const HighlightingRule* singleLineRuleToApply = nullptr;
-        matchedLength = 0;
-
-        for (const HighlightingRule &rule : m_rules) {
-            if (rule.type != RuleType::SingleLine) continue;
-            QRegularExpressionMatchIterator matchIterator = rule.pattern.globalMatch(text);
-             while (matchIterator.hasNext()) {
-                QRegularExpressionMatch match = matchIterator.next();
-                if (match.capturedStart() >= searchStartIndex && match.capturedStart() < earliestMatchStart) {
-                     earliestMatchStart = match.capturedStart();
-                     singleLineRuleToApply = &rule;
-                     matchedLength = match.capturedLength();
-                } else if (match.capturedStart() > earliestMatchStart) {
-                    // Since globalMatch iterates, if we've passed earliestMatchStart, later matches for this rule are past the candidate.
-                    // However, other rules might match earlier.
-                    break;
+        } else if (ruleDef.contains("patterns")) { // List of literal string patterns (e.g., keywords)
+            QJsonArray patternsArray = ruleDef.value("patterns").toArray();
+            for (const QJsonValue& val : patternsArray) {
+                HighlightingRule specificRule = currentRuleSetup; // Copy format
+                specificRule.isBlockRule = false; // Ensure it's not marked as block
+                // For keywords, typically want whole word match, case-sensitive by default
+                QString patternString = "\\b" + QRegularExpression::escape(val.toString()) + "\\b";
+                specificRule.pattern = QRegularExpression(patternString);
+                if (specificRule.pattern.isValid()) {
+                    m_highlightingRules.append(specificRule);
+                } else {
+                    qWarning() << "SyntaxHighlighter: Invalid regex from keyword" << val.toString();
                 }
             }
         }
-
-        if (singleLineRuleToApply) {
-            setFormat(earliestMatchStart, matchedLength, singleLineRuleToApply->format);
-            searchStartIndex = earliestMatchStart + matchedLength;
-        } else {
-            // No more rules match in the remainder of the block
-            searchStartIndex = text.length();
-        }
-         // If a multi-line comment was started in this loop, its state would have been set.
-         // Otherwise, it remains 0 or what previousBlockState dictated if not changed.
-        if (blockStateToSet != 0) setCurrentBlockState(blockStateToSet);
-        else if (previousStateId == 0) setCurrentBlockState(0); // Ensure reset if no multi-line active
     }
 }
 
+void SyntaxHighlighter::highlightBlock(const QString &text) {
+    // Apply non-block rules first
+    for (const HighlightingRule &rule : m_highlightingRules) {
+        if (rule.isBlockRule) continue;
 
-const HighlightingRule* AlteSyntaxHighlighter::findMultiLineRuleByStateId(int stateId) const {
-    for (const auto& rule : m_rules) {
-        if (rule.type == RuleType::MultiLine && rule.stateId == stateId) {
-            return &rule;
+        QRegularExpressionMatchIterator matchIterator = rule.pattern.globalMatch(text);
+        while (matchIterator.hasNext()) {
+            QRegularExpressionMatch match = matchIterator.next();
+            setFormat(match.capturedStart(), match.capturedLength(), rule.format);
         }
     }
-    return nullptr;
+
+    // Handle block rules (multi-line comments, strings etc.)
+    // This state machine is simplified and assumes block rules don't nest in complex ways
+    // with other block rules. It processes one block type at a time per line.
+
+    int blockRuleIndex = previousBlockState(); // 0 = no state, >0 means rule index + 1
+
+    if (blockRuleIndex > 0) { // Resuming a block from previous line
+        int ruleIdx = blockRuleIndex - 1;
+        if (ruleIdx < m_highlightingRules.size() && m_highlightingRules[ruleIdx].isBlockRule) {
+            const HighlightingRule& currentBlockRule = m_highlightingRules[ruleIdx];
+            QRegularExpressionMatch endMatch = currentBlockRule.endPattern.match(text, 0);
+            int endIndex = endMatch.capturedStart();
+            int length;
+
+            if (endIndex == -1) { // Block does not end on this line
+                setCurrentBlockState(blockRuleIndex); // Continue this block state
+                length = text.length();
+                setFormat(0, length, currentBlockRule.format);
+                return; // Entire line is part of this block
+            } else { // Block ends on this line
+                length = endIndex + endMatch.capturedLength();
+                setFormat(0, length, currentBlockRule.format);
+                setCurrentBlockState(0); // Block ended
+                // Now, search for new blocks *after* this one ended
+                // This recursive call is one way, or loop with adjusted startIndex
+                // For simplicity, we'll let the next loop handle new blocks.
+                // highlightBlock(text.mid(length)); // This is not how QSyntaxHighlighter works.
+                // Instead, we need to continue scanning from 'length'.
+                // The logic below will scan from 0, which is fine if blocks don't overlap.
+                // A more robust solution might need a loop here that advances a startIndex.
+            }
+        } else {
+             setCurrentBlockState(0); // Invalid previous state
+        }
+    }
+
+    // Search for new block starts
+    for (int i = 0; i < m_highlightingRules.size(); ++i) {
+        const HighlightingRule &rule = m_highlightingRules[i];
+        if (!rule.isBlockRule) continue;
+
+        QRegularExpressionMatchIterator matchIterator = rule.pattern.globalMatch(text);
+        while (matchIterator.hasNext()) {
+            QRegularExpressionMatch startMatch = matchIterator.next();
+
+            // Check if this match is already formatted by a higher-priority rule or a completed block
+            // This is a simplification; proper handling of overlapping rules is complex.
+            // For now, we assume if currentBlockState became 0, we can format.
+            // A more robust check would involve checking formats at startMatch.capturedStart().
+            if (currentBlockState() != 0 && startMatch.capturedStart() == 0) {
+                 // This part of the line is already consumed by a continued block.
+                 // This check is imperfect. A better way is to manage startIndex.
+                continue;
+            }
+
+
+            QRegularExpressionMatch endMatch = rule.endPattern.match(text, startMatch.capturedEnd());
+            int length;
+            if (endMatch.capturedStart() == -1) { // Block starts here and does not end on this line
+                setCurrentBlockState(i + 1); // Store (rule index + 1) as block state
+                length = text.length() - startMatch.capturedStart();
+                setFormat(startMatch.capturedStart(), length, rule.format);
+                return; // Rest of the line is this block
+            } else { // Block starts and ends on this line
+                length = endMatch.capturedEnd() - startMatch.capturedStart();
+                setFormat(startMatch.capturedStart(), length, rule.format);
+                // Continue searching for more blocks after this one on the same line
+                // This basic loop doesn't handle that well. It will find subsequent blocks
+                // but might re-apply over earlier block matches if not careful.
+                // For now, let's assume one block type takes precedence or they don't overlap this way.
+            }
+        }
+    }
 }
