+++ src/AlteThemeManager.cpp
